See commit for details.
This commit addresses many bugs and changes how the pool is implemented.
We have separated the allocated memory into a stack of pointers, and a
stack of free objects which are constructed. So the difference is that
the stack of allocated space is just that, allocated space. No
constructed objects. It is until we make call like push() that we pop an
allocated space from the stack, that we construct an object and push it
into the free object stack. This enables us to resize(), reserve(),
copy, swap() the pool more easily and safer.